% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{cost.thresh}
\alias{cost.thresh}
\title{Cost/Benefit Optimal threshold}
\usage{
cost.thresh(prob, obs, cost = matrix(c(0, 0, 0, 0), nrow = 2))
}
\arguments{
\item{prob}{Predicted probabilities by a classifier}

\item{obs}{ground truth (correct) 0-1 labels vector}
}
\description{
Compute cost/benefit optimal classification thresholh. See \code{evThreshold}
}
