% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Filter.Vars}
\alias{Filter.Vars}
\title{Filter variables in a data frame}
\usage{
Filter.Vars(dat, rhs.vars, thresh = 0.00016)
}
\arguments{
\item{dat}{data frame}

\item{rhs.vars}{variables in data frame to filter}

\item{threh}{threshold}
}
\description{
Remove constant variable, or rare factor variables below a threshold. Reduce factor levels
}
