% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getSummary}
\alias{getSummary}
\title{sanitaized perfomance results}
\usage{
getSummary(tab, groups = NULL, alpha = 0.05)
}
\description{
This funtion takes a data frame with performance results say from bootstrap of CV with some grouping variable 
and compute mean, and confidence intervals (by groups)
}
