% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{HmscData}
\alias{HmscData}
\title{Hmisc summary data}
\usage{
HmscData(x, digits, prn = any(n != N), pctdig = 0, what = c("\%",
  "proportion"), npct = c("numerator", "both", "denominator", "none"),
  exclude1 = TRUE, vnames = c("labels", "names"), prUnits = TRUE,
  sep = "/", abbreviate.dimnames = FALSE,
  prefix.width = max(nchar(lab)), min.colwidth, formatArgs = NULL,
  round = NULL, prtest = c("P", "stat", "df", "name"), prmsd = FALSE,
  long = FALSE, pdig = 3, eps = 0.001, ...)
}
\arguments{
\item{x}{summary.formula object (reverse). see}

\item{digits}{number of descimal digits. See \code{summary.formula} for the other parameters}
}
\description{
This function converts the Hmisc summary (reverse) object to a dataframe
}
