% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Performance.measures}
\alias{Performance.measures}
\title{Performance Measures}
\usage{
Performance.measures(pred, obs, threshold = NULL)
}
\arguments{
\item{pred}{prediction scores from classifier}

\item{obs}{ground truth  (correct) 0-1 labels vector}
}
\value{
data frame with performance metrics
}
\description{
Compute several performance metrics.
}
