% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{Performance}
\alias{Performance}
\title{Performance metrics}
\usage{
Performance(object, method = "glm", positive.class = "Yes",
  cost = NULL)
}
\arguments{
\item{method}{classifier name}

\item{cost}{costs/benefits matrix  see \code{evThreshold}}

\item{oject}{oject of class trainPipeline}
}
\description{
This function computes several performance metrics from the trained ML pipeline object 
and optionally represent the results in a publication ready tabular format.
}
