% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{reduce.factor}
\alias{reduce.factor}
\title{Reduce Factor levels}
\usage{
reduce.factor(x, old.levels, new.level = NULL)
}
\arguments{
\item{x}{factor variable}

\item{old.levels}{levels to replace}

\item{new.levels}{new level to replace the old levels}
}
\description{
Reduce the levels in a factor variable. Can be use to collapse rare factor levels to a single level
}
