% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynData.R
\name{SynData}
\alias{SynData}
\title{Synthetic Data}
\usage{
SynData(n = 100, p = 10, sim = "simple")
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{sim}{type of simulation: sim = "simple" the two classes are simple multivariate normal distributiobs, 
sim = "hyper", a single multivariate distribution with a noisy hyperplane separating the two classes.}
}
\description{
Generate Synthetic binary classification data from multivariate normal distribution
}
