% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{trainMLpipeline}
\alias{trainMLpipeline}
\alias{trainMLpipeline.default}
\alias{trainMLpipeline.formula}
\title{Train machine learning pipeline models}
\usage{
trainMLpipeline(x, ...)

\method{trainMLpipeline}{default}(x, y, method = NULL,
  trControl = NULL, tuneLength = NULL, positive.class = "Yes",
  best.perf.metric = "ROC", prallel = FALSE, ...)

\method{trainMLpipeline}{formula}(form, data, method = c("glm", "rf",
  "gbm"), trControl = NULL, tuneLength = NULL,
  positive.class = "Yes", best.perf.metric = "ROC", sampling = NULL,
  IMR = 5, parallel = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link{UnsupRF}}}

\item{\dots}{further arguments passed to or from other methods.}

\item{data}{data.frame or matrix}

\item{RFdist}{RF distance matrix computed from \code{\link{RFdist}}.}

\item{B}{number of bootstraps}

\item{clustermethod}{clustering method, options are \code{pamkCBI}, or \code{claraCBI}, or \code{hclustCBI}. 
  Not to sure about \code{hclustCBI} see the \code{fpc} package. \code{pamkCBI} is 
  recommended for RF dissimilarity matrix, but we have found standard 
\code{hclust} in base R works well with  Ward's minimum variance creterion}

\item{classification}{type of prediction for finding optimal number of clusters 
see \code{\link[fpc]{nselectboot}}.}

\item{krange}{integer vector; numbers of clusters to be tried}

\item{kopt}{user provided optimal number of clusters}

\item{run.boot}{(logical) run bootstrap cluster-wise stability ?}

\item{fun}{function to determine mediods, should be \code{mean}, \code{median}, 
or \code{sum}. See \code{\link{mediod}}}
}
\value{
A list with elements:  
\enumerate{
\item cluster.model: The cluster model  
\item cluster: cluster memberships 
\item kopt: optimal number of clusters  
\item mediods: a mediod object            
}
}
\description{
This function trains the machine learni ng pipeline framework. Specifically, its simply 
provides an easy means to train several machine learning models through the \code{caret} \code{train} function
By default (i.e setting the method parameter to NULL), the pipeline trains 9 models and select the best 
model for making final predictions
}
