% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{VimPlot}
\alias{VimPlot}
\title{Plot variable importance}
\usage{
VimPlot(x, top = min(20, length(x$importance)), ...)
}
\arguments{
\item{x}{one column data frame with variable importance. Row names corresponds to the variables}

\item{to}{plot the top variables}
}
\description{
plot variable importance for random forest
}
